moved LineBreaksRenderer under utils folder
