+++ src/utils/LineBreaksRenderer.js
